Add proper JavaScript errors instead of strings in contract/method.js
A lot of the contract methods returned strings as errors instead of
error objects. This made it hard to debug errors as no stack trace or
additional information was available. I've replaced all errors with a
custom error class which allows attaching extra data to errors, for
example, the concerned transaction. This makes it easier for developers
to debug issues.